ukvm: Refactor header files, add checked overflow macros
