--- ukvm/processor-flags.h
@@ -1,50 +1,0 @@
-#ifndef _ASM_X86_PROCESSOR_FLAGS_H
-#define _ASM_X86_PROCESSOR_FLAGS_H
-/* Various flags defined: can be included from assembler. */
-
-#ifndef _BITUL
-
-#ifdef __ASSEMBLY__
-#define _AC(X,Y)	X
-#define _AT(T,X)	X
-#else
-#define __AC(X,Y)	(X##Y)
-#define _AC(X,Y)	__AC(X,Y)
-#define _AT(T,X)	((T)(X))
-#endif
-
-#define _BITUL(x)	(_AC(1,UL) << (x))
-#define _BITULL(x)	(_AC(1,ULL) << (x))
-
-#endif
-
-/*
- * EFLAGS bits
- */
-#define X86_EFLAGS_CF	0x00000001 /* Carry Flag */
-
-/*
- * Basic CPU control in CR0
- */
-#define X86_CR0_PE_BIT		0 /* Protection Enable */
-#define X86_CR0_PE		_BITUL(X86_CR0_PE_BIT)
-#define X86_CR0_PG_BIT		31 /* Paging */
-#define X86_CR0_PG		_BITUL(X86_CR0_PG_BIT)
-
-/*
- * Intel CPU features in CR4
- */
-#define X86_CR4_PAE_BIT		5 /* enable physical address extensions */
-#define X86_CR4_PAE		_BITUL(X86_CR4_PAE_BIT)
-
-/*
- * Intel long mode page directory/table entries
- */
-#define X86_PDPT_P_BIT          0 /* Present */
-#define X86_PDPT_P              _BITUL(X86_PDPT_P_BIT)
-#define X86_PDPT_RW_BIT         1 /* Writable */
-#define X86_PDPT_RW             _BITUL(X86_PDPT_RW_BIT)
-#define X86_PDPT_PS_BIT         7 /* Page size */
-#define X86_PDPT_PS             _BITUL(X86_PDPT_PS_BIT)
-
-#endif
