+++ ukvm/ukvm-blk.c
@@ -8,8 +8,9 @@
 #include <unistd.h>
 #include <assert.h>
 
+#include "ukvm-private.h"
+#include "ukvm-modules.h"
 #include "ukvm.h"
-#include "ukvm_modules.h"
 
 static struct ukvm_blkinfo blkinfo;
 static char *diskfile;
